+++ app/assets/javascripts/views/shared/chartWidgetView.js
@@ -26,7 +26,9 @@
       // Inner width of the chart, used internally
       _width: null,
       // Inner height of the chart, used internally
-      _height: null
+      _height: null,
+      // Flag that determines whether the widget is visible or not
+      visible: true
     },
 
     initialize: function (settings) {
@@ -193,7 +195,7 @@
           }
           this.chart = chart({
             el: this.chartContainer,
-            // By using the SVG renderer, we give the client the posibility to
+            // By using the SVG renderer, we give the client the possibility to
             // translate the text contained in the charts
             renderer: 'svg'
           }).update();
@@ -204,7 +206,7 @@
 
       // We save the state of the widget each time we render as it can be the
       // consequence of a change in the configuration
-      // NOTE: We need to make sure in case the view hasn't been instanciated with
+      // NOTE: We need to make sure in case the view hasn't been instantiated with
       // a chart configuration that it's set by the toolbox before triggering
       this.trigger('state:change', {
         type: 'chart',
@@ -212,7 +214,8 @@
         x: this.options.columnX,
         y: this.options.columnY,
         xLabel: this.options.xLabel,
-        yLabel: this.options.yLabel
+        yLabel: this.options.yLabel,
+        visible: this.options.visible
       });
     },
 
@@ -273,7 +276,32 @@
       // We append the button to the DOM
       this.el.querySelector('.js-switch-button').appendChild(button);
     },
+    /**
+     * Render the Toggle Visibility button and attach an event listener to it
+     */
+    _renderToggleVisibilityButton: function () {
+      // We create the button
+      var button = document.createElement('button');
+      button.type = 'button';
+      button.classList.add('c-button', 'toggle-visibility-button');
+      if (!this.options.visible) button.classList.add('-slashed');
+      button.textContent = 'Hide';
+      // We attach the listener
+      button.addEventListener('click', function() {
+        button.classList.toggle('-slashed');
+        this._toggleVisibility();
+      }.bind(this));
 
+      // We append the button to the DOM
+      this.el.querySelector('.js-toggle-visibility-button').appendChild(button);
+    },
+    /**
+     * Trigger widget visibility change
+     */
+    _toggleVisibility: function () {
+      this.options.visible = !this.options.visible;
+      this.trigger('state:change', { visible: this.options.visible });
+    },
     /**
      * Render the Custom Axis Label Inputs and attach an event listener to it
      */
@@ -325,6 +353,7 @@
         }
         if (this.options.switchCallback && typeof this.options.switchCallback === 'function') {
           this._renderSwitchButton();
+          this._renderToggleVisibilityButton();
         }
       }